requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==4.2
django-cors-headers==3.13.0
django-debug-toolbar==4.0.0
django-filter==24.3
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.3
psycopg==3.1.14
psycopg-binary==3.1.14
PyJWT==2.9.0
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
